Create Make a Circle
Suppose you are in an art class now. The teacher wants you to make a circle, made up of cloth, with string glued on the edge. Given the radius of the circle, you should calculate the area of the cloth and the length of the string that you need in order to make this circle.

You should use macro to define PI as 3.14f.