Create EAN-13 Submissions
Numbering the positions from the right (exclude check digit), the odd positions in the twelve-digit number are weight of 3 and the even positions are weight of 1. Multiply each data digit with its corresponding weight.

For example:

data: 4 7 1 0 3 6 7 6 6 3 2 7

weight: 1 3 1 3 1 3 1 3 1 3 1 3

Calculate the sum of the results from step 1.

Subtract the sum from the nearest multiple of 10 that is equal to or higher than the sum (see hint). The result is the check digit.

Write a program that calculate the check digit.