Create Caesar Cipher Encoder
The Caesar cipher (a cipher is a method of encrypting data) is one of the simplest and most widely known encryption techniques. For English text, the Caesar cipher would work by taking each letter in the plaintext message and substituting the letter that is k letters latter (allowing wraparound; that is, having the letter z followed by the letter a) in the alphabet.

For example, if k = 3, then the letter a in plaintext becomes d in ciphertext, B in plaintext becomes E in ciphertext, and the plaintext message "bob, i love you. Alice" becomes "ere, l oryh brx. Dolfh" in ciphertext. Note that k may be negative. It means that you should substitute the letter in the plaintext message that is |k| letters earlier.

Write a program that prints out the Caesar cipher for the given input (described in Input Format).

Note: Transform only English alphabet (case-preserving) and preserve other characters such as Arabic numerals or punctuation.